pyproject.toml
[project]
name = "filetype-detector"
version = "0.1.0"
description = "A flexible Python library for file type inference using multiple strategies including extension-based, magic number-based, and AI-powered detection"
authors = [
    { name = "devcomfort", email = "im@devcomfort.me" }
]
license = { text = "MIT" }
dependencies = [
    "python-magic>=0.4.27",
    "magika>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.8"
keywords = ["file-type", "mime-type", "file-detection", "magic-numbers", "ai-detection"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.2",
    "loguru>=0.7.3",
]
docs = [
    "mkdocs>=1.6.1",
    "mkdocstrings[python]>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "build>=1.3.0",
    "twine>=6.2.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/filetype_detector"]
